pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cex-arb-bot"
version = "0.1.0"
description = "Cross-exchange arbitrage bot for detecting and executing price gaps between Binance, OKX, and other exchanges"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Arbitrage Bot Developer", email = "dev@example.com"}
]
keywords = ["arbitrage", "trading", "cryptocurrency", "binance", "okx"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

dependencies = [
    "ccxt>=4.0.0",
    "ccxt[pro]>=4.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "orjson>=3.9.0",
    "tenacity>=8.2.0",
    "loguru>=0.7.0",
    "python-telegram-bot>=20.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "aiosqlite>=0.19.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pyarrow>=14.0.0",
    "uvloop>=0.17.0; sys_platform != 'win32'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[project.scripts]
cex-arb = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
